pyproject.toml
[project]
name = "fds-aws-coding-exercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.67",
    "python-dotenv>=1.2.1",
]
